clean up cedrusdb wrapper code
